Simplified promote to newsroom #32
 * Removes the use of 'Promote to frontpage' for featuring in the
 newsroom.
 * Adds a checkbox to (be published) news that can add it directly into
 the existing entity reference field for listing featured items.
 * Changes view to show entity referenced featured items in the featured
 display, and exclude them from the other news.
 * Removes custom views override and logic requried to combine the
 'promoted to frontpage' with the entity reference featured.

Highlights that will all need reviewing in revisions and scheduling is
introduced. For the use case now gives editors an easy way to add
directly to featured as the 'Promote to frontpage' did, while removing
the complexity of having to maintain and combine the two different
methods of selecting articles.